Removed polling station detail endpoint, refactored code, however one test fails, plz help
