+++ frontend/app/module/input/PollingStation/PollingStationHomePage.tsx
@@ -1,6 +1,6 @@
 import { Link } from "react-router-dom";
 
-export function DataEntryIndexPage() {
+export function PollingStationHomePage() {
   return (
     <span>
       Klik <Link to={"./recounted"}>hier</Link> om met de invoer te beginnen.
