+++ frontend/app/module/input/PollingStation/index.ts
@@ -1,7 +1,6 @@
 export * from "./CandidatesVotesPage";
-export * from "./DataEntryIndexPage";
+export * from "./PollingStationHomePage.tsx";
 export * from "./DifferencesPage";
-export * from "./DataEntryLayout";
 export * from "./PollingStationLayout";
 export * from "./RecountedPage";
 export * from "./VotersAndVotesPage";
