+++ frontend/app/module/input/page/InputHomePage.test.tsx
@@ -3,12 +3,12 @@ import { describe, expect, test } from "vitest";
 import { overrideOnce, render, screen } from "app/test/unit";
 
 import { ElectionProvider } from "@kiesraad/api";
-import { electionMock } from "@kiesraad/api-mocks";
+import { electionMock, electionMockResponse } from "@kiesraad/api-mocks";
 
 import { InputHomePage } from "./InputHomePage";
 
 describe("InputHomePage", () => {
-  overrideOnce("get", "/api/elections/1", 200, { election: electionMock });
+  overrideOnce("get", "/api/elections/1", 200, electionMockResponse);
   test("Election name", async () => {
     render(
       <ElectionProvider electionId={1}>
