+++ frontend/lib/api-mocks/PollingStationMockData.ts
@@ -1,8 +1,4 @@
-import {
-  PollingStation,
-  PollingStationDetailsResponse,
-  PollingStationListResponse,
-} from "@kiesraad/api";
+import { PollingStation, PollingStationListResponse } from "@kiesraad/api";
 
 export const pollingStationMockData: PollingStation = {
   election_id: 1,
@@ -16,13 +12,19 @@ export const pollingStationMockData: PollingStation = {
   street: "Rijksweg A12",
 };
 
-export const pollingStationDetailsMockResponse: PollingStationDetailsResponse = {
-  polling_station: pollingStationMockData,
-};
-
 export const pollingStationListMockResponse: PollingStationListResponse = {
   polling_stations: [
-    pollingStationMockData,
+    {
+      election_id: 1,
+      id: 1,
+      number: 33,
+      name: 'Stembureau "Op Rolletjes"',
+      house_number: "1",
+      locality: "Den Haag",
+      polling_station_type: "Mobiel",
+      postal_code: "1234 YQ",
+      street: "Rijksweg A12",
+    },
     {
       election_id: 1,
       id: 2,
