+++ frontend/lib/api/index.ts
@@ -10,5 +10,4 @@ export * from "./useApiPostRequest";
 export * from "./useElectionDataRequest";
 export * from "./useElectionListRequest";
 export * from "./usePollingStationDataEntry";
-export * from "./usePollingStationDataRequest";
 export * from "./usePollingStationListRequest";
