+++ frontend/lib/api/polling_station/index.ts
@@ -1,4 +1,2 @@
 export * from "./PollingStationListProvider";
-export * from "./PollingStationProvider";
-export * from "./usePollingStation";
 export * from "./usePollingStationList";
